#11 Adds separate maintenance mode by given env. variable
Previously the nginx dogu regularly checked the CES registry in order
to switch to a maintenance page. This behaviour is no longer possible
during a restore of a CES backup because the registry is no longer
available.

For this use case a there are two things to accomplish in order to
provide a maintenance page:

1. The original nginx dogu must be stopped
1. A new maintenance nginx dogu must be created with the environment
variable `CES_MAINTENANCE_MODE` set to `true`.

That new nginx dogu does not connect to the registry but rather shows a
static page during the restore. As the maintenance nginx dogu is 100 %
based on the same nginx dogu and the existing image can be re-used. Once
the restore was done, the maintenance nginx must be properly disposed so
the original nginx can be started.