#11 Refactor environment variable for proper usage
The startup scripts does not really need to be provided with a
environment variable because it is already available globally (i.e. the
environment).

The Dockerfile entrypoint is fixed in this regard, so the call does not
need another argument.

The startup script itself is changed in that way that it reads the env.
variable instead of using a program argument.